+++ src/python/DAS/__init__.py
@@ -7,9 +7,5 @@ DAS modules
 
 .. moduleauthor:: Valentin Kuznetsov <vkuznet@gmail.com>
 """
-
-__revision__ = "$Id: __init__.py,v 1.8 2010/04/30 17:14:35 valya Exp $"
-__version__ = "$Revision: 1.8 $"
 __author__ = "Valentin Kuznetsov"
-
-version = "0.5.7"
+version = "0.5.8"