+++ src/python/DAS/utils/jsonwrapper/__init__.py
@@ -2,26 +2,26 @@
 #-*- coding: ISO-8859-1 -*-
 
 """
-JSON wrapper around different JSON python implementation.
-So far we use simplejson (json) and cjson, other modules can be
-added in addition.
+JSON wrapper around different JSON python implementations.
+We use simplejson (json), cjson and yajl JSON implementation.
 """
 
-__revision__ = "$Id: __init__.py,v 1.3 2010/01/04 19:01:21 valya Exp $"
-__version__ = "$Revision: 1.3 $"
-__author__ = "Valentin Kuznetsov"
-__all__ = ['loads', 'dumps', 'JSONEncoder']
+__author__ = "Valentin Kuznetsov <vkuznet@gmail.com>"
 
+_module = None
 try:
-    import cjson
+    import yajl
+    _module = "yajl"
 except:
-    pass
-try:
-    import json # python 2.6 and later
-except:
-    import simplejson as json # python 2.5 and earlier
+    try:
+        import cjson
+        _module = "cjson"
+    except:
+        pass
 
-_module = "cjson"
+import json
+if  not _module: # use default JSON module
+    _module = "json"
 
 def loads(idict, **kwargs):
     """
@@ -31,6 +31,10 @@ def loads(idict, **kwargs):
         return json.loads(idict, **kwargs)
     elif _module == 'cjson':
         return cjson.decode(idict)
+    elif _module == 'yajl':
+        return yajl.loads(idict, **kwargs)
+    else:
+        raise Exception("Not support JSON module: %s" % _module)
 
 def load(source):
     """
@@ -39,12 +43,15 @@ def load(source):
     objects.
     """
     return json.load(source)
-
-def JSONEncoder(*args, **kwargs):
-    """
-    This class is only available in non-cjson
-    """
-    return json.JSONEncoder(*args, **kwargs)
+    if  _module == 'json':
+        return json.load(source)
+    elif _module == 'cjson':
+        data = source.read()
+        return cjson.decode(data)
+    elif _module == 'yajl':
+        return yajl.load(source)
+    else:
+        raise Exception("Not support JSON module: %s" % _module)
 
 def dumps(idict, **kwargs):
     """
@@ -53,14 +60,62 @@ def dumps(idict, **kwargs):
     if  _module == 'json':
         return json.dumps(idict, **kwargs)
     elif _module == 'cjson':
-        try:
-            data = cjson.encode(idict)
-            return data
-        except:
-            try: 
-                data = json.dumps(idict, **kwargs)
-                return data
-            except:
-                msg = 'Unable to ecode %s' % idict
-                raise Exception(msg)
+        return cjson.encode(idict)
+    elif _module == 'yajl':
+        return yajl.dumps(idict, **kwargs)
+    else:
+        raise Exception("Not support JSON module: %s" % _module)
+
+def dump(doc, source):
+    """
+    Use json.dump for back-ward compatibility, since cjson doesn't
+    provide this method. The dump method works on file-descriptor
+    objects.
+    """
+    if  _module == 'json':
+        return json.dump(doc, source)
+    elif _module == 'cjson':
+        stj = cjson.encode(doc)
+        return source.write(stj)
+    elif _module == 'yajl':
+        return yajl.dump(doc, source)
+    else:
+        raise Exception("Not support JSON module: %s" % _module)
+
+class JSONEncoder(object):
+    """
+    JSONEncoder wrapper
+    """
+    def __init__(self, **kwargs):
+        self.encoder = json.JSONEncoder(kwargs)
+        self.kwargs = kwargs
+
+    def encode(self, idict):
+        """Decode JSON method"""
+        if  _module == 'cjson':
+            return cjson.encode(idict)
+        elif _module == 'yajl':
+            return yajl.Encoder(self.kwargs).encode(idict)
+        return self.encoder.encode(idict)
+
+    def iterencode(self, idict):
+        return self.encoder.iterencode(idict)
+
+class JSONDecoder(object):
+    """
+    JSONDecoder wrapper
+    """
+    def __init__(self, **kwargs):
+        self.decoder = json.JSONDecoder(kwargs)
+
+    def decode(self, idict):
+        """Decode JSON method"""
+        if  _module == 'cjson':
+            return cjson.decode(idict)
+        elif _module == 'yajl':
+            return jajl.Decoder(self.kwargs).decode(idict)
+        return self.decoder.decode(idict)
+
+    def raw_decode(self, idict):
+        return self.decoder.raw_decode(idict)